Extract Crucible body renderers into module
Moved Crucible body-focused logic out of `crucibleChrome.tsx` into new `crucibleBodyRenderers.tsx`, including semantic item generation, canvas/SVG body rendering, bond drawing, and shared description/number helpers. Updated `CrucibleChart`, server physics chart configs, and related tests to import from the new module. This refactor keeps behavior intact while separating chrome/projection concerns from body rendering logic for cleaner reuse and maintenance.
